Some nighttime optimizations

+++ functions/Get-PowerUpPackage.ps1
@@ -24,36 +24,52 @@ Function Get-PowerUpPackage {
 			ValueFromPipeline = $true)]
 		[Alias('FileName', 'Name', 'Package')]
 		[string[]]$Path,
-		[string[]]$Build
+		[string[]]$Build,
+		[switch]$Unpacked
 	)
 	begin {
 
 	}
 	process {
 		foreach ($pFile in (Get-Item $Path)) {
-			#Create temp folder
-			$workFolder = New-TempWorkspaceFolder
+			if (!$Unpacked) {
+				#Create temp folder
+				$workFolder = New-TempWorkspaceFolder
+			}
+			else {
+				$workFolder = $pFile
+			}
 
 			#Create output object and set values for default fields
 			$packageInfo = $pFile
 
-			Add-Member -InputObject $packageInfo -MemberType AliasProperty -Name Size -Value Length
+			if (!$Unpacked) {
+				Add-Member -InputObject $packageInfo -MemberType AliasProperty -Name Size -Value Length
+			}
 			Add-Member -InputObject $packageInfo -MemberType AliasProperty -Name Path -Value FullName
 			
 
 
 			try {
-				#Extract package file
-				Write-Verbose "Extracting package file from the archive $pFile to $workFolder"
-				Expand-ArchiveItem -Path $pFile -DestinationPath $workFolder -Item "PowerUp.package.json"
+				if (!$Unpacked) {
+					#Extract package file
+					Write-Verbose "Extracting package file from the archive $pFile to $workFolder"
+					Expand-ArchiveItem -Path $pFile -DestinationPath $workFolder -Item "PowerUp.package.json"
+				}
 
 				#Load package object
 				Write-Verbose "Loading package information from $pFile"
 				$package = [PowerUpPackage]::FromFile((Join-Path $workFolder "PowerUp.package.json"))
+				$moduleManifest = 'Modules\PowerUp\PowerUp.psd1'
 
-				#Extract config file
-				Write-Verbose "Extracting config file $($package.ConfigurationFile) from the archive $pFile to $workFolder"
-				Expand-ArchiveItem -Path $pFile -DestinationPath $workFolder -Item $package.ConfigurationFile
+				if (!$Unpacked) {
+					#Extract config and module files
+					Write-Verbose "Extracting config file $($package.ConfigurationFile) from the archive $pFile to $workFolder"
+					Expand-ArchiveItem -Path $pFile -DestinationPath $workFolder -Item $package.ConfigurationFile
+
+					Write-Verbose "Extracting module manifest from the archive $pFile to $workFolder"
+					Expand-ArchiveItem -Path $pFile -DestinationPath $workFolder -Item $moduleManifest
+				}
 
 				#Load configuration 
 				$configPath = Join-Path $workFolder $package.ConfigurationFile
@@ -63,7 +79,8 @@ Function Get-PowerUpPackage {
 				$config = Get-PowerUpConfig $configPath
 				Add-Member -InputObject $packageInfo -MemberType NoteProperty -Name Config -Value $config
 
-				Add-Member -InputObject $packageInfo -MemberType NoteProperty -Name Version -Value $package.Builds[-1].Build
+				Add-Member -InputObject $packageInfo -MemberType NoteProperty -Name Version -Value $package.GetVersion()
+				Add-Member -InputObject $packageInfo -MemberType NoteProperty -Name ModuleVersion -Value (Test-ModuleManifest (Join-Path $workFolder $moduleManifest)).Version
 				
 				#Generate build and script objects
 				$builds = @()
@@ -90,7 +107,7 @@ Function Get-PowerUpPackage {
 				throw $_
 			}
 			finally {
-				if ($workFolder.Name -like 'PowerUpWorkspace*') {
+				if (!$Unpacked -and $workFolder.Name -like 'PowerUpWorkspace*') {
 					Remove-Item $workFolder -Recurse -Force
 				}
 			}
